refactor: extract homepage state persistence into custom hook
- Create `useHomepageState` hook to encapsulate state management
- Extract state persistence logic to `homepage-state-service.ts`
- Improve code readability and reusability as suggested in review
- Handles Map serialization for localStorage correctly
- Maintains same functionality with cleaner architecture
